Stage 1/6 Addition Refactored
